Replace "AI" German translation
The previous translation is full of translations that make no sense in the context of Geometry Dash and oddities which just aren't how folks actually speak. In other words, it's provably just a DeepL translation, copy pasted, with little to no adjustments made.

This version is for the most part made from scratch by paw and fixes those issues. It does contain some odd language, but that's intentional flavor, in accordance with prominent Anglosaxon online culture.